Troquei os double do preço por float
Só postei esse separado para lembrar de não deixar os códigos parecidos com os do professor sem necessidade

+++ API_Jogos/src/main/java/com/jogos/api/dto/GameDTO.java
@@ -10,7 +10,7 @@
         private String developer;
         private String distributor;
         private int score;
-        private Double price;
+        private float price;
         private String genre;
         private int rating;
         private boolean hasDLC;
@@ -47,11 +47,11 @@
             this.distributor = distributor;
         }
 
-        public Double getPrice() {
+        public float getPrice() {
             return price;
         }
 
-        public void setPrice(Double price) {
+        public void setPrice(float price) {
             this.price = price;
         }
